requirements.txt
transformers
Pillow
pypdf
jq
langchain==0.0.330
streamlit
boto3
awscli
bs4
setuptools
PyPDF2
pypdf
langchain_community
python-dotenv
faiss-cpu